pyproject.toml
[tool.poetry]
name = "blur-faces"
version = "0.1.0"
description = ""
authors = ["*Vinicius_Goia* <*vinicius@naturalengines.com*>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0.80"
imutils = "^0.5.4"
numpy = "^1.26.4"
streamlit = "^1.33.0"
scipy = "^1.13.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
